+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_freeradiusclients/__init__.py
@@ -1,0 +1,4 @@
+"""FreeRadius Clients plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .freeradiusclients import Plugin  # NOQA