+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_freeradiuseapconf/__init__.py
@@ -1,0 +1,4 @@
+"""FreeRadius AP Config plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .freeradiuseapconf import Plugin  # NOQA