+++ src/netgate_xml_to_xlsx/plugins/plugin_installed_freeradiusinterfaces/__init__.py
@@ -1,0 +1,4 @@
+"""FreeRadius Interfaces plugin."""
+# Copyright © 2022 Appropriate Solutions, Inc. All rights reserved.
+
+from .freeradiusinterfaces import Plugin  # NOQA