+++ src/netgate_xml_to_xlsx/plugins/plugin_ipsec/ipsec.py
@@ -162,7 +162,7 @@ class Plugin(BasePlugin):
         if client_node is None:
             return
         if self.should_process(client_node):
-            print(f"WARNING: {self.node_ancesters(client_node)} is unimplemented.")
+            print(f"WARNING: {self.node_path(client_node)} is unimplemented.")
 
         return []